Revert "LPS-123758 Move classes that are not exported to internal package - portal-mobile-device-detection-fiftyonedegrees"
This reverts commit afd691658591ef4ab036c82d24236be674ff00ef.
